+++ lib/two_stage_pipeliner/visualizers/core/images_data.py
