+++ lib/two_stage_pipeliner/visualizers/pipeline.py
@@ -4,20 +4,20 @@ from IPython.display import display
 from two_stage_pipeliner.core.data import ImageData
 from two_stage_pipeliner.core.visualizer import Visualizer
 from two_stage_pipeliner.core.batch_generator import BatchGeneratorImageData
-from two_stage_pipeliner.inferencers.detection import DetectionInferencer
+from two_stage_pipeliner.inferencers.pipeline import PipelineInferencer
 from two_stage_pipeliner.utils.jupyter_visualizer import JupyterVisualizer
-from two_stage_pipeliner.visualization.core.images_data import visualize_image_data, \
+from two_stage_pipeliner.visualizers.core.images_data import visualize_image_data, \
     visualize_images_data_side_by_side
 
 
-class DetectionVisualizer(Visualizer):
-    def __init__(self, inferencer: DetectionInferencer = None):
+class PipelineVisualizer(Visualizer):
+    def __init__(self, inferencer: PipelineInferencer = None):
         Visualizer.__init__(self, inferencer)
         self.jupyter_visualizer = None
 
     def visualize(self,
                   images_data: List[ImageData],
-                  score_threshold: float = None):
+                  detection_score_threshold: float = None):
         image_names = [image_data.image_path.name for image_data in images_data]
         images_data_gen = BatchGeneratorImageData(images_data, batch_size=1)
 
@@ -26,15 +26,15 @@ class DetectionVisualizer(Visualizer):
             if self.inferencer is None:
                 display(visualize_image_data(
                     true_image_data,
-                    use_labels=False,
+                    use_labels=True,
                     score_type=None
                 ))
             else:
-                pred_image_data = self.inferencer.predict([batch], score_threshold)[0]
+                pred_image_data = self.inferencer.predict([batch], detection_score_threshold)[0]
                 display(visualize_images_data_side_by_side(
                     true_image_data, pred_image_data,
-                    use_labels1=False, use_labels2=False,
-                    score_type1=None, score_type2='detection'
+                    use_labels1=True, use_labels2=True,
+                    score_type1=None, score_type2=None
                 ))
 
         if self.jupyter_visualizer is not None:
